Fix issue #42
