+++ src/Download/Download.js
@@ -10,46 +10,10 @@ const Download = () => {
     const [videoPlaying, setVideoPlaying] = useState(false);
     const [currentImageIndex, setCurrentImageIndex] = useState(0);
 
-    const toggleMute = () => {
-        setIsMuted(!isMuted);
-        if (audioRef.current) {
-            audioRef.current.volume = isMuted ? 0.5 : 0; // Set volume to 1 if currently muted, 0 if currently unmuted
-            if (isMuted) {
-                audioRef.current.play().catch(err => {
-                    console.error("Error playing audio:", err);
-                });
-            } else {
-                audioRef.current.pause();   
-            }
-        }
-    };
-
-    useEffect(() => {
-        toggleMute();
-        const handleLocationChange = () => {
-            setIsMuted(true); 
-            if (audioRef.current) {
-                audioRef.current.pause();
-            }
-        };
-    }, []);
     
-    useEffect(() => {
-        const videoElement = document.querySelector('video');
-        if (videoElement) {
-            videoElement.onplay = () => {
-                setVideoPlaying(true);
-            };
-            videoElement.onerror = () => {
-                setVideoPlaying(false);
-            };
-        }
-        if (audioRef.current) {
-            audioRef.current.play().catch(err => {
-                console.error("Error playing audio:", err);
-            });
-        }
-    }, []);
+
+  
+    
 
 
     return (
@@ -88,21 +52,9 @@ const Download = () => {
                 </div>
             </div>
 
-            <button onClick={toggleMute} className="fixed bottom-4 right-4 bg-red-600 hover:bg-red-800 text-white font-bold py-2 px-4 rounded flex items-center justify-center rounded">
-            {isMuted ? (
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="w-6 h-6 stroke-black">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="m9 9 10.5-3m0 6.553v3.75a2.25 2.25 0 01-1.632 2.163l-1.32.377a1.803 1.803 0 01-.99-3.467l2.31-.66a2.25 2.25 0 001.632-2.163Zm0 0V2.25L9 5.25v10.303m0 0v3.75a2.25 2.25 0 01-1.632 2.163l-1.32.377a1.803 1.803 0 01-.99-3.467l2.31-.66A2.25 2.25 0 009 15.553Z" />
-                        <path strokeLinecap="round" d="M3 3l18 18" />
-                    </svg>
-                ) : (
+                
 
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="w-6 h-6 stroke-black">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="m9 9 10.5-3m0 6.553v3.75a2.25 2.25 0 01-1.632 2.163l-1.32.377a1.803 1.803 0 01-.99-3.467l2.31-.66a2.25 2.25 0 001.632-2.163Zm0 0V2.25L9 5.25v10.303m0 0v3.75a2.25 2.25 0 01-1.632 2.163l-1.32.377a1.803 1.803 0 01-.99-3.467l2.31-.66A2.25 2.25 0 009 15.553Z" />
-                    </svg>
-                )}
-            </button>            
-
-            <audio ref={audioRef} src="main_menu.mp3" loop />
+            
         </div>
     )
 }
